Fix frontpage hamburger menu on short phones
Previously the font size of the hamburger menu was to large, so it would
get cut of before exposing all entries. This makes the font smaller to
fit on shorter phones.

The issue of the user being able to scroll the webpage behind the menu
has also been taken care of.